feat: add simple auth

+++ src/routes/postgre.ts
@@ -1,6 +1,7 @@
 import express from 'express';
 import multer from 'multer';
-import { deleteAndPublishLayer, deleteLayer } from 'service/geoserver';
+import { verify } from 'service/auth';
+import { deleteAndPublishLayer, deleteLayer, publishLayer } from 'service/geoserver';
 import { addRecordGeoJSON, addRecordGeoJSONFromMongoQuery } from 'service/postgre';
 import { readSHPFile } from 'service/shapefile';
 import { unzipFile } from 'service/unzipper';
@@ -17,6 +18,13 @@ router.post('/ShapefileImport', upload.fields([{
   name: 'dbf', maxCount: 1
 }]), async function (req, res) {
   const body: any = req.body
+
+  const authStatus = await verify(body?.token);
+  if (authStatus?.status == 403) {
+    res.send(authStatus)
+    return;
+  }
+
   const files = req.files as { [fieldname: string]: Express.Multer.File[] };
   if (!files.shp?.[0]?.buffer) {
     res.status(200).send({
@@ -39,6 +47,13 @@ router.post('/ShapefileImport', upload.fields([{
 
 router.post('/ShapefileZipImport', upload.single('shapefile_zip'), async function (req, res) {
   const body: any = req.body
+
+  const authStatus = await verify(body?.token);
+  if (authStatus?.status == 403) {
+    res.send(authStatus)
+    return;
+  }
+
   const file = req.file as Express.Multer.File
   let files = await unzipFile(file.buffer)
   const fileData = await readSHPFile(files.shp, files.dbf)
@@ -56,11 +71,18 @@ router.post('/ShapefileZipImport', upload.single('shapefile_zip'), async functio
 
 router.post('/DataFromMongoQueryImport', async function (req, res) {
   const body: any = req.body
+
+  const authStatus = await verify(body?.token);
+  if (authStatus?.status == 403) {
+    res.send(authStatus)
+    return;
+  }
+
   if (!body?.tableNameImport || !body?.layerName || !body?.db || !body?.collection) {
     res.status(400).send("db, collection, layerName, tableNameImport is required")
     return
   }
-  await addRecordGeoJSONFromMongoQuery(
+  let numberRecordAdded = await addRecordGeoJSONFromMongoQuery(
     {
       tableNameImport: body?.tableNameImport,
       layerTitle: body?.layerTitle || body?.layerName,
@@ -74,19 +96,32 @@ router.post('/DataFromMongoQueryImport', async function (req, res) {
       tinh_thanh: body?.tinh_thanh,
       clearCacheTinhThanh: body?.clearCacheTinhThanh
     })
-  await deleteAndPublishLayer({
-    host: `${process.env.HOST_GEOSERVER}/geoserver/rest`,
-    workspaceName: 'bando',
-    wmsstoreName: 'geostore',
-    layerName: body?.layerName,
-    tableName: body?.tableNameImport,
-    layerTitle: body?.layerTitle || body?.layerName
-  })
-  res.status(200).send('ok')
+  if (numberRecordAdded) {
+    await deleteAndPublishLayer({
+      host: `${process.env.HOST_GEOSERVER}/geoserver/rest`,
+      workspaceName: 'bando',
+      wmsstoreName: 'geostore',
+      layerName: body?.layerName,
+      tableName: body?.tableNameImport,
+      layerTitle: body?.layerTitle || body?.layerName
+    })
+    res.status(200).send('done')
+  }
+  else {
+    res.status(400).send(`Đã thêm ${numberRecordAdded} bản ghi`)
+  }
+  res.status(200).send()
 })
 
 router.post('/GeoJSONImport', async function (req, res) {
   const body: any = req.body
+
+  const authStatus = await verify(body?.token);
+  if (authStatus?.status == 403) {
+    res.send(authStatus)
+    return;
+  }
+
   const kq = await addRecordGeoJSON(body?.tableNameImport, body?.geoJsonData);
   if (kq) {
     await deleteAndPublishLayer({
@@ -105,13 +140,39 @@ router.post('/GeoJSONImport', async function (req, res) {
 router.post('/LayerRemove', async function (req, res) {
   const body: any = req.body
 
+  const authStatus = await verify(body?.token);
+  if (authStatus?.status == 403) {
+    res.send(authStatus)
+    return;
+  }
+
   const kq = await deleteLayer({
     host: body?.host || `${process.env.HOST_GEOSERVER}/geoserver/rest`,
-    workspaceName: 'bando',
-    wmsstoreName: 'geostore',
+    workspaceName: body?.workspaceName || 'bando',
+    wmsstoreName: body?.wmsstoreName || 'geostore',
     layerName: body?.layerName
   })
-  res.status(200).send(kq)
+  console.log('LayerRemove', body?.layerName, kq);
+
+  res.status(kq?.status || 500).send(kq)
+})
+router.post('/LayerPublish', async function (req, res) {
+  const body: any = req.body
+  const authStatus = await verify(body?.token);
+  if (authStatus?.status == 403) {
+    res.send(authStatus)
+    return;
+  }
+
+  const kq = await publishLayer({
+    host: body?.host || `${process.env.HOST_GEOSERVER}/geoserver/rest`,
+    workspaceName: body?.workspaceName || 'bando',
+    wmsstoreName: body?.wmsstoreName || 'geostore',
+    layerName: body?.layerName,
+    tableName: body?.tableNameImport,
+    layerTitle: body?.layerTitle
+  })
+  res.status(kq?.status || 500).send(kq)
 })
 
 export default router
